Move hoc to filter file so it has context
