+++ assets/js/hocs/index.js
@@ -7,4 +7,3 @@ export { default as withProductVariations } from './with-product-variations';
 export { default as withSearchedProducts } from './with-searched-products';
 export { default as withTransformSingleSelectToMultipleSelect } from './with-transform-single-select-to-multiple-select';
 export { default as withRestApiHydration } from './with-rest-api-hydration';
-export { default as withDefaultAttributes } from './with-default-attributes';
